+++ cmd/root.go
@@ -323,7 +323,7 @@ func Execute() error {
 
 	info, ok := debug.ReadBuildInfo()
 	v, c, d := buildVersionInfo(Version, Commit, Date, info, ok)
-	RootCmd.Version = fmt.Sprintf("%s, commit %s, built at %s", v, c, d)
+	RootCmd.Version = versionLine(v, c, d)
 	return RootCmd.Execute()
 }